pyproject.toml
[project]
name = "agentlab2"
version = "0.1.0"
description = "agentlab2, open source agentic evaluation and data generation framework."
readme = "README.md"
authors = [
    { name = "Aman Jaiswal", email = "amanjaiswal73892@gmail.com" },
    { name = "Oleh Shliazhko", email = "oleh.shliazhko@servicenow.com" },
]
requires-python = ">=3.13"
dependencies = [
    "pydantic~=2.0",
    "litellm~=1.80.8",
    "playwright~=1.49",
    "Pillow~=12.0",
    "beautifulsoup4~=4.14",
    "numpy~=2.3.5",
    "ray[default]>=2.52.1",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
fix = true
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
extend-select = ["I"] # sort imports

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]